+++ src/pages/LeaderboardPage/components/LinkWalletsModal/components/index.ts
@@ -1,0 +1,3 @@
+export * from './LinkedWalletsList'
+export * from './LinkingBlock'
+export * from './VerifyWalletBlock'
